Update u6.py, fix mistake in docstrings
Change the numbers in the docstrings of setDOState() getDIState() and getDIOState().
The docstrings state that arguments ioNum = 20, 21, 22 access the so called MIO digital I/O. The class BitStateRead will take the argument mod 20, s.t. the user intending to read MIO1 reads FIO1 etc. The hardware documentation states that MIO 0-2 are shared with CIO 0-2.